Implement background loading and first go at scrolling refresh view
DOES NOT WORK YET!
