+++ app/src/main/res/layout/frag_main_schedule.xml
@@ -31,25 +31,8 @@
         android:text="@string/blablabla"
         android:textColor="@android:color/black"
         android:textSize="22sp"
-        android:layout_toStartOf="@+id/scheduleCurrentWeek"
         android:layout_toEndOf="@id/calendarImageButton"
-        android:layout_toRightOf="@id/calendarImageButton"
-        android:layout_toLeftOf="@+id/scheduleCurrentWeek" />
-
-    <TextView
-        android:id="@+id/scheduleCurrentWeek"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@id/calendarImageButton"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentTop="true"
-        android:fontFamily="monospace"
-        android:gravity="center_vertical"
-        android:paddingEnd="12dp"
-        android:paddingStart="12dp"
-        android:text="@string/blablabla"
-        android:textSize="18sp"
-        android:layout_alignParentRight="true" />
+        android:layout_toRightOf="@id/calendarImageButton" />
 
     <RelativeLayout
         android:id="@+id/noSchoolView"
